refactor- Alterade name service
