+++ tech-docs-master/src/app/services/pdf-thumbnail.service.ts
